Consolidate NewGateway and Gateway in one module
